Add movie db class

+++ src/main/java/com/example/demo/restservice/Movie.java
@@ -1,16 +1,16 @@
 package com.example.demo.restservice;
 
 public class Movie {
-    private final long id;
+    private final int id;
     private final String  title;
 
 
-    public Movie(long id, String title){
+    public Movie(int id, String title){
         this.id = id;
         this.title = title;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
